ssj-experiment-results/requirements.txt
numpy==1.19.1
pandas==1.1.3
matplotlib==3.4.1
kafka-python==2.0.2
